Fix teardown and callback refcount ABI

+++ Source/C++ Class/DLABCallbackBase.h
@@ -21,17 +21,17 @@ template <typename Derived, typename DelegateT>
 class DLABCallbackBase {
 protected:
     __weak DelegateT delegate;
-    std::atomic<uint32_t> refCount{1};
+    std::atomic<ULONG> refCount{1};
     
 public:
     explicit DLABCallbackBase(DelegateT delegate) : delegate(delegate) {}
     
-    uint32_t AddRef() {
+    ULONG AddRef() {
         return ++refCount;
     }
     
-    uint32_t Release() {
-        uint32_t newRefValue = --refCount;
+    ULONG Release() {
+        ULONG newRefValue = --refCount;
         if (newRefValue == 0) {
             delete static_cast<Derived *>(this);
             return 0;
